more moving to int

+++ src/base/intSize.h
@@ -2,6 +2,6 @@
 #define SRC_BASE_INTSIZE_H_
 
 typedef u_int64_t HashType;
-typedef u_int64_t UInt;
-typedef int64_t Int;
+typedef u_int32_t UInt;
+typedef int32_t Int;
 #endif /* SRC_BASE_INTSIZE_H_ */
